pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-agent-protocol"
version = "0.1.0"
description = "A universal, framework-agnostic LangChain Agent protocol for cross-project portability"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["langchain", "agent", "llm", "ai", "chatbot", "protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "langchain-core>=0.1.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
all = [
    "langchain-agent-protocol[openai,anthropic]",
]

[project.urls]
Homepage = "https://github.com/yourusername/langchain-agent-protocol"
Documentation = "https://langchain-agent-protocol.readthedocs.io"
Repository = "https://github.com/yourusername/langchain-agent-protocol"
"Bug Tracker" = "https://github.com/yourusername/langchain-agent-protocol/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["langchain_agent_protocol*"]
exclude = ["tests*", "examples*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
